+++ spec/models/team_spec.rb
@@ -111,7 +111,7 @@ describe Team do
   end
   context '#inform!' do
     let(:team) { Fabricate(:team) }
-    it 'sends message to all channels', vcr: { cassette_name: 'slack/channels_list' } do
+    it 'sends message to all channels', vcr: { cassette_name: 'slack/conversations_list' } do
       expect_any_instance_of(Slack::Web::Client).to receive(:chat_postMessage).exactly(3).times.and_return('ts' => '1503435956.000247')
       team.inform!(message: 'message')
     end